[Mach-O] Emit __data_in_code

+++ macho/macho.h
@@ -498,7 +498,7 @@ struct SourceVersionCommand {
 
 struct DataInCodeEntry {
   u32 offset;
-  u32 length;
+  u16 length;
   u16 kind;
 };
